sizeof

+++ 0x00-hello_world/6-size.c
@@ -9,10 +9,10 @@
 int main(void)
 {
 
-	printf("Size of an integer: %lu byte(s)\n", sizeof(int));
-	printf("Size of a float: %lu byte(s)\n", sizeof(float));
-	printf("Size of characters: %lu byte(s)\n", sizeof(char));
-	printf("Size of long integer: %lu byte(s)\n", sizeof(long int));
-	printf("Size of long long integer: %lu byte(s)", sizeof(long long int));
+	printf("Size of an integer: %lu byte(s)\n", (unsigned long)sizeof(int));
+	printf("Size of a float: %lu byte(s)\n", (unsigned long)sizeof(float));
+	printf("Size of characters: %lu byte(s)\n", (unsigned long)sizeof(char));
+	printf("Size of long integer: %lu byte(s)\n", (unsigned long)sizeof(long int));
+	printf("Size of long long integer: %lu byte(s)", (unsigned long)sizeof(long long int));
 	return (0);
 }
